# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail app Api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "deendipo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://retail-app.test/api",
            "description": " API Documentation Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authenticate user",
                "description": "Authenticate user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "bolean",
                                        "example": "mailer@gmail2.com"
                                    },
                                    "password": {
                                        "type": "bolean",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "bolean",
                                            "example": "Login successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "4"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Olamilekan"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Adeniyi"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "mailer@gmail2.com"
                                                },
                                                "user_type": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "email_verified_at": {
                                                    "type": "null",
                                                    "example": "null"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcmV0YWlsLWFwcC50ZXN0L2FwaS9hdXRoL2xvZ2luIiwiaWF0IjoxNjU3NjA5NjMwLCJleHAiOjE2NTc2MTMyMzAsIm5iZiI6MTY1NzYwOTYzMCwianRpIjoiZ1E3ZUNGMjZRamJ2VEVqUiIsInN1YiI6IjQiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.IcfCIE-Nopl-IOzK3aIABa-WYZVtt991S3LUsGxSpIw"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error."
                    },
                    "405": {
                        "description": "Bad method."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register auth",
                "description": "Register auth",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Cabana"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Copa"
                                    },
                                    "email": {
                                        "type": "bolean",
                                        "example": "mailer@gmail22.com"
                                    },
                                    "password": {
                                        "type": "bolean",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "bolean",
                                            "example": "Registration successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "4"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Cabana"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Copa"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "mailer@gmail22.com"
                                                },
                                                "user_type": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "email_verified_at": {
                                                    "type": "null",
                                                    "example": "null"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcmV0YWlsLWFwcC50ZXN0L2FwaS9hdXRoL2xvZ2luIiwiaWF0IjoxNjU3NjA5NjMwLCJleHAiOjE2NTc2MTMyMzAsIm5iZiI6MTY1NzYwOTYzMCwianRpIjoiZ1E3ZUNGMjZRamJ2VEVqUiIsInN1YiI6IjQiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.IcfCIE-Nopl-IOzK3aIABa-WYZVtt991S3LUsGxSpIw"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error."
                    },
                    "401": {
                        "description": "Bad method."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}